feat: 관리자 페이지 및 틀린그림찾기 문제 등록 기능 추가
- admin-frontend: 퀴즈 문제 관리 페이지 구현
  - 이미지 클릭으로 틀린 부분 좌표 설정
  - S3 업로드 최적화 (제출 시에만 업로드)
  - 반지름 및 허용 오차 조정 기능
  - 실시간 미리보기 (빨간 원 표시)

- backend/services/admin: 관리자 서비스 추가
  - 퀴즈 문제 CRUD API
  - S3 이미지 업로드 지원
  - tolerance 필드 정수(픽셀) 타입으로 처리

- backend/services/quiz: tolerance 타입 수정
  - DB: INTEGER 타입 사용
  - 모델: sql.NullInt32로 변경
  - validator: int32 파라미터로 변경

- docker-compose: admin-service 추가

+++ backend/services/quiz/internal/service/quiz_service.go
@@ -185,7 +185,7 @@ func (s *quizServiceImpl) SubmitAnswer(ctx context.Context, userID string, quest
 		isCorrect = s.validator.ValidateRegionSelect(
 			rsAnswer.SelectedRegion,
 			question.CorrectRegions,
-			question.Tolerance.Int32,
+			int32(question.Tolerance.Int32),
 		)
 
 	case repository.QuestionTypeComparison:
